Fixed slow tests.
This addresses issue #21 and is related to issue #18.

Eliminated the slow testGame(), as it's no longer needed. However this was only
a symptom of the underlying problems. Loading the configuration file is very
slow, as demonstrated by issue #18. The Trek class was also initialising member
variables too soon, which slowed down some tests. Other tests were slow due to
not being isolated from cooperating classes. This was especially true for the
long range scanner tests; see also issue #32.